+++ src/main/java/com/commercehub/sensu/api/Client.java
@@ -1,0 +1,52 @@
+package com.commercehub.sensu.api;
+
+import java.util.List;
+
+public class Client {
+    private String name;
+    private String address;
+    private List<String> subscriptions;
+    private int timestamp;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public List<String> getSubscriptions() {
+        return subscriptions;
+    }
+
+    public void setSubscriptions(List<String> subscriptions) {
+        this.subscriptions = subscriptions;
+    }
+
+    public int getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(int timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    @Override
+    public String toString() {
+        return "Client{" +
+                "name='" + name + '\'' +
+                ", address='" + address + '\'' +
+                ", subscriptions=" + subscriptions +
+                ", timestamp=" + timestamp +
+                '}';
+    }
+}